+++ ogi-parent/ogi-services/src/main/java/fr/jerep6/ogi/persistance/bo/RealPropertyBuilt.java
@@ -45,7 +45,7 @@ import fr.jerep6.ogi.utils.DocumentUtils;
 @Setter
 public abstract class RealPropertyBuilt extends RealProperty {
 	@Column(name = "PRB_AREA")
-	private Integer			area;
+	private Float			area;
 
 	@Column(name = "PRB_BUILD_DATE")
 	@Temporal(TemporalType.DATE)
@@ -88,10 +88,10 @@ public abstract class RealPropertyBuilt extends RealProperty {
 	// table of the owning entity but you can override them using@AttributeOverride:
 	@Embedded
 	@AttributeOverrides({ //
-	@AttributeOverride(name = "kWh", column = @Column(name = "PRB_DPE_KWH")), //
-			@AttributeOverride(name = "classificationKWh", column = @Column(name = "PRB_DPE_CLASS_KWH")), //
-			@AttributeOverride(name = "ges", column = @Column(name = "PRB_DPE_GES")), //
-			@AttributeOverride(name = "classificationGes", column = @Column(name = "PRB_DPE_CLASS_GES")), //
+		@AttributeOverride(name = "kWh", column = @Column(name = "PRB_DPE_KWH")), //
+		@AttributeOverride(name = "classificationKWh", column = @Column(name = "PRB_DPE_CLASS_KWH")), //
+		@AttributeOverride(name = "ges", column = @Column(name = "PRB_DPE_GES")), //
+		@AttributeOverride(name = "classificationGes", column = @Column(name = "PRB_DPE_CLASS_GES")), //
 	})
 	private DPE				dpe;
